Move the JDT shim code out of _ext and into lib-extra, compiling against p2deps.
